+++ zcash_client_backend/src/data_api/wallet.rs
@@ -1,5 +1,6 @@
 use std::num::NonZeroU32;
 
+use nonempty::NonEmpty;
 use rand_core::OsRng;
 use sapling::{
     note_encryption::{try_sapling_note_decryption, PreparedIncomingViewingKey},
@@ -10,10 +11,10 @@ use zcash_keys::{
     keys::{HDSeedFingerprint, UnifiedAddressRequest, UnifiedFullViewingKey},
 };
 use zcash_primitives::{
-    consensus::{self, NetworkUpgrade},
+    consensus::{self, BlockHeight, NetworkUpgrade},
     memo::MemoBytes,
     transaction::{
-        builder::{BuildConfig, Builder},
+        builder::{BuildConfig, BuildResult, Builder},
         components::amount::{Amount, NonNegativeAmount},
         fees::{zip317::FeeError as Zip317FeeError, FeeRule, StandardFeeRule},
         Transaction, TxId,
@@ -25,12 +26,14 @@ use zip32::DiversifierIndex;
 use crate::{
     address::Address,
     data_api::{
-        error::Error, wallet::input_selection::Proposal, DecryptedTransaction, SentTransaction,
-        SentTransactionOutput, WalletCommitmentTrees, WalletRead, WalletWrite,
+        error::Error, DecryptedTransaction, SentTransaction, SentTransactionOutput,
+        WalletCommitmentTrees, WalletRead, WalletWrite,
     },
     decrypt_transaction,
     fees::{self, DustOutputPolicy},
     keys::UnifiedSpendingKey,
+    proposal::ProposalError,
+    proposal::{self, Proposal},
     wallet::{Note, OvkPolicy, Recipient},
     zip321::{self, Payment},
     PoolType, ShieldedProtocol,
@@ -45,9 +48,12 @@ use super::InputSource;
 
 #[cfg(feature = "transparent-inputs")]
 use {
-    crate::wallet::WalletTransparentOutput, input_selection::ShieldingSelector,
-    sapling::keys::OutgoingViewingKey, std::convert::Infallible,
-    zcash_keys::encoding::AddressCodec, zcash_primitives::legacy::TransparentAddress,
+    input_selection::ShieldingSelector,
+    sapling::keys::OutgoingViewingKey,
+    std::convert::Infallible,
+    zcash_keys::encoding::AddressCodec,
+    zcash_primitives::legacy::TransparentAddress,
+    zcash_primitives::transaction::components::{OutPoint, TxOut},
 };
 
 /// Represents the identifier for an account that was derived from a ZIP-32 HD seed and account index.
@@ -133,11 +139,11 @@ where
 }
 
 #[allow(clippy::needless_doctest_main)]
-/// Creates a transaction paying the specified address from the given account.
+/// Creates a transaction or series of transactions paying the specified address from
+/// the given account, and the [`TxId`] corresponding to each newly-created transaction.
 ///
-/// Returns the row index of the newly-created transaction in the `transactions` table
-/// within the data database. The caller can read the raw transaction bytes from the `raw`
-/// column in order to broadcast the transaction to the network.
+/// These transactions can be retrieved from the underlying data store using the
+/// [`WalletRead::get_transaction`] method.
 ///
 /// Do not call this multiple times in parallel, or you will generate transactions that
 /// double-spend the same notes.
@@ -263,7 +269,7 @@ pub fn create_spend_to_address<DbT, ParamsT>(
     min_confirmations: NonZeroU32,
     change_memo: Option<MemoBytes>,
 ) -> Result<
-    TxId,
+    NonEmpty<TxId>,
     Error<
         <DbT as WalletRead>::Error,
         <DbT as WalletCommitmentTrees>::Error,
@@ -296,7 +302,7 @@ where
         change_memo,
     )?;
 
-    create_proposed_transaction(
+    create_proposed_transactions(
         wallet_db,
         params,
         spend_prover,
@@ -307,15 +313,12 @@ where
     )
 }
 
-/// Constructs a transaction that sends funds as specified by the `request` argument
-/// and stores it to the wallet's "sent transactions" data store, and returns a
-/// unique identifier for the transaction; this identifier is used only for internal
-/// reference purposes and is not the same as the transaction's txid, although after v4
-/// transactions have been made invalid in a future network upgrade, the txid could
-/// potentially be used for this type (as it is non-malleable for v5+ transactions).
+/// Constructs a transaction or series of transactions that send funds as specified
+/// by the `request` argument, stores them to the wallet's "sent transactions" data
+/// store, and returns the [`TxId`] for each transaction constructed.
 ///
-/// This procedure uses the wallet's underlying note selection algorithm to choose
-/// inputs of sufficient value to satisfy the request, if possible.
+/// The newly-created transactions can be retrieved from the underlying data store using the
+/// [`WalletRead::get_transaction`] method.
 ///
 /// Do not call this multiple times in parallel, or you will generate transactions that
 /// double-spend the same notes.
@@ -374,7 +377,7 @@ pub fn spend<DbT, ParamsT, InputsT>(
     ovk_policy: OvkPolicy,
     min_confirmations: NonZeroU32,
 ) -> Result<
-    TxId,
+    NonEmpty<TxId>,
     Error<
         <DbT as WalletRead>::Error,
         <DbT as WalletCommitmentTrees>::Error,
@@ -404,7 +407,7 @@ where
         min_confirmations,
     )?;
 
-    create_proposed_transaction(
+    create_proposed_transactions(
         wallet_db,
         params,
         spend_prover,
@@ -415,8 +418,8 @@ where
     )
 }
 
-/// Select transaction inputs, compute fees, and construct a proposal for a transaction
-/// that can then be authorized and made ready for submission to the network with
+/// Select transaction inputs, compute fees, and construct a proposal for a transaction or series
+/// of transactions that can then be authorized and made ready for submission to the network with
 /// [`create_proposed_transaction`].
 #[allow(clippy::too_many_arguments)]
 #[allow(clippy::type_complexity)]
@@ -459,9 +462,13 @@ where
         .map_err(Error::from)
 }
 
-/// Proposes a transaction paying the specified address from the given account.
+/// Proposes making a payment to the specified address from the given account.
 ///
-/// Returns the proposal, which may then be executed using [`create_proposed_transaction`]
+/// Returns the proposal, which may then be executed using [`create_proposed_transaction`].
+/// Depending upon the recipient address, more than one transaction may be constructed
+/// in the execution of the returned proposal.
+///
+/// This method uses the basic [`GreedyInputSelector`] for input selection.
 ///
 /// Parameters:
 /// * `wallet_db`: A read/write reference to the wallet database.
@@ -531,6 +538,8 @@ where
     )
 }
 
+/// Constructs a proposal to shield all of the funds belonging to the provided set of
+/// addresses.
 #[cfg(feature = "transparent-inputs")]
 #[allow(clippy::too_many_arguments)]
 #[allow(clippy::type_complexity)]
@@ -572,17 +581,20 @@ where
         .map_err(Error::from)
 }
 
-/// Construct, prove, and sign a transaction using the inputs supplied by the given proposal,
-/// and persist it to the wallet database.
+/// Construct, prove, and sign a transaction or series of transactions using the inputs supplied by
+/// the given proposal, and persist it to the wallet database.
 ///
-/// Returns the database identifier for the newly constructed transaction, or an error if
+/// Returns the database identifier for each newly constructed transaction, or an error if
 /// an error occurs in transaction construction, proving, or signing.
 ///
-/// Note: If the payment includes a recipient with an Orchard-only UA, this will attempt
-/// to fall back to the transparent receiver until full Orchard support is implemented.
+/// When evaluating multi-step proposals, only transparent outputs of any given step may be spent
+/// in later steps; attempting to spend a shielded note (including change) output by an earlier
+/// step is not supported, because the ultimate positions of those notes in the global note
+/// commitment tree cannot be known until the transaction that produces those notes is mined,
+/// and therefore the required spend proofs for such notes cannot be constructed.
 #[allow(clippy::too_many_arguments)]
 #[allow(clippy::type_complexity)]
-pub fn create_proposed_transaction<DbT, ParamsT, InputsErrT, FeeRuleT, N>(
+pub fn create_proposed_transactions<DbT, ParamsT, InputsErrT, FeeRuleT, N>(
     wallet_db: &mut DbT,
     params: &ParamsT,
     spend_prover: &impl SpendProver,
@@ -591,7 +603,7 @@ pub fn create_proposed_transaction<DbT, ParamsT, InputsErrT, FeeRuleT, N>(
     ovk_policy: OvkPolicy,
     proposal: &Proposal<FeeRuleT, N>,
 ) -> Result<
-    TxId,
+    NonEmpty<TxId>,
     Error<
         <DbT as WalletRead>::Error,
         <DbT as WalletCommitmentTrees>::Error,
@@ -604,6 +616,92 @@ where
     ParamsT: consensus::Parameters + Clone,
     FeeRuleT: FeeRule,
 {
+    let mut step_results = Vec::with_capacity(proposal.steps().len());
+    for step in proposal.steps() {
+        let step_result = create_proposed_transaction(
+            wallet_db,
+            params,
+            spend_prover,
+            output_prover,
+            usk,
+            ovk_policy.clone(),
+            proposal.fee_rule(),
+            proposal.min_target_height(),
+            &step_results,
+            step,
+        )?;
+        step_results.push((step, step_result));
+    }
+
+    Ok(NonEmpty::from_vec(
+        step_results
+            .iter()
+            .map(|(_, r)| r.transaction().txid())
+            .collect(),
+    )
+    .expect("proposal.steps is NonEmpty"))
+}
+
+#[allow(clippy::too_many_arguments)]
+#[allow(clippy::type_complexity)]
+fn create_proposed_transaction<DbT, ParamsT, InputsErrT, FeeRuleT, N>(
+    wallet_db: &mut DbT,
+    params: &ParamsT,
+    spend_prover: &impl SpendProver,
+    output_prover: &impl OutputProver,
+    usk: &UnifiedSpendingKey,
+    ovk_policy: OvkPolicy,
+    fee_rule: &FeeRuleT,
+    min_target_height: BlockHeight,
+    prior_step_results: &[(&proposal::Step<N>, BuildResult)],
+    proposal_step: &proposal::Step<N>,
+) -> Result<
+    BuildResult,
+    Error<
+        <DbT as WalletRead>::Error,
+        <DbT as WalletCommitmentTrees>::Error,
+        InputsErrT,
+        FeeRuleT::Error,
+    >,
+>
+where
+    DbT: WalletWrite + WalletCommitmentTrees,
+    ParamsT: consensus::Parameters + Clone,
+    FeeRuleT: FeeRule,
+{
+    // TODO: Spending shielded outputs of prior multi-step transaction steps is not yet
+    // supported. Maybe support this at some point? Doing so would require a higher-level
+    // approach in the wallet that waits for transactions with shielded outputs to be
+    // mined and only then attempts to perform the next step.
+    for s_ref in proposal_step.prior_step_inputs() {
+        prior_step_results.get(s_ref.step_index()).map_or_else(
+            || {
+                // Return an error in case the step index doesn't match up with a step
+                Err(Error::Proposal(ProposalError::ReferenceError(*s_ref)))
+            },
+            |step| match s_ref.output_index() {
+                proposal::StepOutputIndex::Payment(i) => {
+                    let prior_pool = step
+                        .0
+                        .payment_pools()
+                        .get(&i)
+                        .ok_or(Error::Proposal(ProposalError::ReferenceError(*s_ref)))?;
+
+                    if matches!(prior_pool, PoolType::Shielded(_)) {
+                        Err(Error::ProposalNotSupported)
+                    } else {
+                        Ok(())
+                    }
+                }
+                proposal::StepOutputIndex::Change(_) => {
+                    // Only shielded change is supported by zcash_client_backend, so multi-step
+                    // transactions cannot yet spend prior transactions' change outputs.
+                    Err(Error::ProposalNotSupported)
+                }
+            },
+        )?;
+    }
+
     let account = wallet_db
         .get_account_for_ufvk(&usk.to_unified_full_viewing_key())
         .map_err(Error::DataSource)?
@@ -620,7 +718,7 @@ where
 
     let internal_ovk = || {
         #[cfg(feature = "transparent-inputs")]
-        return if proposal.is_shielding() {
+        return if proposal_step.is_shielding() {
             Some(OutgoingViewingKey(
                 usk.transparent()
                     .to_account_pubkey()
@@ -635,7 +733,7 @@ where
         Some(dfvk.to_ovk(Scope::Internal))
     };
 
-    let (sapling_anchor, sapling_inputs) = proposal.shielded_inputs().map_or_else(
+    let (sapling_anchor, sapling_inputs) = proposal_step.shielded_inputs().map_or_else(
         || Ok((sapling::Anchor::empty_tree(), vec![])),
         |inputs| {
             wallet_db.with_sapling_tree_mut::<_, _, Error<_, _, _, _>>(|sapling_tree| {
@@ -680,7 +778,7 @@ where
     // are no possible transparent inputs, so we ignore those
     let mut builder = Builder::new(
         params.clone(),
-        proposal.min_target_height(),
+        min_target_height,
         BuildConfig::Standard {
             sapling_anchor: Some(sapling_anchor),
             orchard_anchor: None,
@@ -692,38 +790,98 @@ where
     }
 
     #[cfg(feature = "transparent-inputs")]
-    let utxos = {
+    let utxos_spent = {
         let known_addrs = wallet_db
             .get_transparent_receivers(account)
             .map_err(Error::DataSource)?;
 
-        let mut utxos: Vec<WalletTransparentOutput> = vec![];
-        for utxo in proposal.transparent_inputs() {
-            utxos.push(utxo.clone());
-
+        let mut utxos_spent: Vec<OutPoint> = vec![];
+        let mut add_transparent_input = |addr: &TransparentAddress,
+                                         outpoint: OutPoint,
+                                         utxo: TxOut|
+         -> Result<
+            (),
+            Error<
+                <DbT as WalletRead>::Error,
+                <DbT as WalletCommitmentTrees>::Error,
+                InputsErrT,
+                FeeRuleT::Error,
+            >,
+        > {
             let address_metadata = known_addrs
-                .get(utxo.recipient_address())
-                .ok_or_else(|| Error::AddressNotRecognized(*utxo.recipient_address()))?
+                .get(addr)
+                .ok_or(Error::AddressNotRecognized(*addr))?
                 .clone()
-                .ok_or_else(|| Error::NoSpendingKey(utxo.recipient_address().encode(params)))?;
+                .ok_or_else(|| Error::NoSpendingKey(addr.encode(params)))?;
 
             let secret_key = usk
                 .transparent()
-                .derive_external_secret_key(address_metadata.address_index())
+                .derive_secret_key(address_metadata.scope(), address_metadata.address_index())
                 .unwrap();
 
-            builder.add_transparent_input(
-                secret_key,
+            utxos_spent.push(outpoint.clone());
+            builder.add_transparent_input(secret_key, outpoint, utxo)?;
+
+            Ok(())
+        };
+
+        for utxo in proposal_step.transparent_inputs() {
+            add_transparent_input(
+                utxo.recipient_address(),
                 utxo.outpoint().clone(),
                 utxo.txout().clone(),
             )?;
         }
-        utxos
+        for input_ref in proposal_step.prior_step_inputs() {
+            match input_ref.output_index() {
+                proposal::StepOutputIndex::Payment(i) => {
+                    // We know based upon the earlier check that this must be a transparent input,
+                    // We also know that transparent outputs for that previous step were added to
+                    // the transaction in payment index order, so we can use dead reckoning to
+                    // figure out which output it ended up being.
+                    let (prior_step, result) = &prior_step_results[input_ref.step_index()];
+                    let recipient_address = match &prior_step
+                        .transaction_request()
+                        .payments()
+                        .get(&i)
+                        .expect("Payment step references are checked at construction")
+                        .recipient_address
+                    {
+                        Address::Transparent(t) => Some(t),
+                        Address::Unified(uaddr) => uaddr.transparent(),
+                        _ => None,
+                    }
+                    .ok_or(Error::ProposalNotSupported)?;
+                    let outpoint = OutPoint::new(
+                        result.transaction().txid().into(),
+                        u32::try_from(
+                            prior_step
+                                .payment_pools()
+                                .iter()
+                                .filter(|(_, pool)| pool == &&PoolType::Transparent)
+                                .take_while(|(j, _)| j <= &&i)
+                                .count()
+                                - 1,
+                        )
+                        .expect("Transparent output index fits into a u32"),
+                    );
+                    let utxo = &result
+                        .transaction()
+                        .transparent_bundle()
+                        .ok_or(Error::Proposal(ProposalError::ReferenceError(*input_ref)))?
+                        .vout[outpoint.n() as usize];
+
+                    add_transparent_input(recipient_address, outpoint, utxo.clone())?;
+                }
+                proposal::StepOutputIndex::Change(_) => unreachable!(),
+            }
+        }
+        utxos_spent
     };
 
     let mut sapling_output_meta = vec![];
     let mut transparent_output_meta = vec![];
-    for payment in proposal.transaction_request().payments() {
+    for payment in proposal_step.transaction_request().payments().values() {
         match &payment.recipient_address {
             Address::Unified(ua) => {
                 let memo = payment
@@ -772,12 +930,12 @@ where
                 } else {
                     builder.add_transparent_output(to, payment.amount)?;
                 }
-                transparent_output_meta.push((*to, payment.amount));
+                transparent_output_meta.push((to, payment.amount));
             }
         }
     }
 
-    for change_value in proposal.balance().proposed_change() {
+    for change_value in proposal_step.balance().proposed_change() {
         let memo = change_value
             .memo()
             .map_or_else(MemoBytes::empty, |m| m.clone());
@@ -812,7 +970,7 @@ where
     }
 
     // Build the transaction with the specified fee rule
-    let build_result = builder.build(OsRng, spend_prover, output_prover, proposal.fee_rule())?;
+    let build_result = builder.build(OsRng, spend_prover, output_prover, fee_rule)?;
 
     let internal_ivk = PreparedIncomingViewingKey::new(&dfvk.to_ivk(Scope::Internal));
     let sapling_outputs =
@@ -837,10 +995,7 @@ where
                             try_sapling_note_decryption(
                                 &internal_ivk,
                                 &bundle.shielded_outputs()[output_index],
-                                consensus::sapling_zip212_enforcement(
-                                    params,
-                                    proposal.min_target_height(),
-                                ),
+                                consensus::sapling_zip212_enforcement(params, min_target_height),
                             )
                             .map(|(note, _, _)| (account, note))
                         })
@@ -867,7 +1022,7 @@ where
 
         SentTransactionOutput::from_parts(
             output_index,
-            Recipient::Transparent(addr),
+            Recipient::Transparent(*addr),
             value,
             None,
             None,
@@ -880,22 +1035,22 @@ where
             created: time::OffsetDateTime::now_utc(),
             account,
             outputs: sapling_outputs.chain(transparent_outputs).collect(),
-            fee_amount: Amount::from(proposal.balance().fee_required()),
+            fee_amount: Amount::from(proposal_step.balance().fee_required()),
             #[cfg(feature = "transparent-inputs")]
-            utxos_spent: utxos.iter().map(|utxo| utxo.outpoint().clone()).collect(),
+            utxos_spent,
         })
         .map_err(Error::DataSource)?;
 
-    Ok(build_result.transaction().txid())
+    Ok(build_result)
 }
 
-/// Constructs a transaction that consumes available transparent UTXOs belonging to
-/// the specified secret key, and sends them to the default address for the provided Sapling
-/// extended full viewing key.
+/// Constructs a transaction that consumes available transparent UTXOs belonging to the specified
+/// secret key, and sends them to the most-preferred receiver of the default internal address for
+/// the provided Unified Spending Key.
 ///
 /// This procedure will not attempt to shield transparent funds if the total amount being shielded
-/// is less than the default fee to send the transaction. Fees will be paid only from the transparent
-/// UTXOs being consumed.
+/// is less than the default fee to send the transaction. Fees will be paid only from the
+/// transparent UTXOs being consumed.
 ///
 /// Parameters:
 /// * `wallet_db`: A read/write reference to the wallet database
@@ -936,7 +1091,7 @@ pub fn shield_transparent_funds<DbT, ParamsT, InputsT>(
     from_addrs: &[TransparentAddress],
     min_confirmations: u32,
 ) -> Result<
-    TxId,
+    NonEmpty<TxId>,
     Error<
         <DbT as WalletRead>::Error,
         <DbT as WalletCommitmentTrees>::Error,
@@ -958,7 +1113,7 @@ where
         min_confirmations,
     )?;
 
-    create_proposed_transaction(
+    create_proposed_transactions(
         wallet_db,
         params,
         spend_prover,